Added checking of docker socket.

+++ ws
@@ -4,9 +4,22 @@ set -e
 
 command -v docker >/dev/null 2>&1 || { echo >&2 "I require docker but it's not installed.  Aborting."; exit 1; }
 
+docker_sock_volume=""
+# Check if we have docker on host
+if [ -S "/var/run/docker.sock" ]; then
+  docker_sock_volume="--volume=/var/run/docker.sock:/var/run/docker.sock"
+fi
+
+# Prepare ssh keys
+if [ ! -d "$HOME/.ssh" ]; then
+   printf ${red}"You need to setup your ssh keys first."${neutral}"\n"
+   exit 1
+fi
+
 docker run --rm -ti \
     -v $(pwd):/var/www/app \
     -v $HOME/.ssh:/home/daker/.ssh \
     -v $HOME/.docker-workspace:/home/daker/.docker-workspace \
+    $docker_sock_volume \
     --env PGID=$(id -g) --env PUID=$(id -u) \
     travissouth/workspace $@
